@@ -8,6 +8,7 @@ import com.snowgears.shop.shop.ComboShop;
 import com.snowgears.shop.shop.ShopType;
 import com.snowgears.shop.util.DisplayUtil;
 import com.snowgears.shop.util.ItemListType;
+import com.snowgears.shop.util.ShopLogger;
 import com.snowgears.shop.util.UtilMethods;
 import org.bukkit.*;
 import org.bukkit.block.Block;
@@ -30,6 +31,8 @@ import org.bukkit.scheduler.BukkitScheduler;
 import java.io.File;
 import java.io.IOException;
 import java.util.*;
+import java.util.AbstractMap.SimpleEntry;
+import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
@@ -57,6 +60,23 @@ public class ShopHandler {
     private ArrayList<ItemStack> itemListItems = new ArrayList<>();
 
     private ArrayList<UUID> playersSavingShops = new ArrayList<>();
+    
+    // Map to track player last processed locations for movement-based display updates
+    private ConcurrentHashMap<UUID, Location> lastProcessedLocations = new ConcurrentHashMap<>();
+
+    // Cache for player connections to avoid expensive reflection calls
+    private ConcurrentHashMap<UUID, Object> playerConnectionCache = new ConcurrentHashMap<>();
+
+    // Default chunk radius for shop location searches and maximum display distance
+    // these values come from the Shop class configuration - see getShopSearchRadius() and getMaxShopDisplayDistance()
+    
+    // Very close distance for immediate shop display
+    private static final double CLOSE_SHOP_DISPLAY_DISTANCE = 2.0;
+
+    // Teleport cooldown map to prevent multiple display updates during teleportation
+    private ConcurrentHashMap<UUID, Long> teleportCooldowns = new ConcurrentHashMap<>();
+    // Cooldown time in milliseconds (500ms = half a second)
+    private static final long TELEPORT_COOLDOWN_MS = 500;
 
     public ShopHandler(Shop instance) {
         plugin = instance;
@@ -64,12 +84,23 @@ public class ShopHandler {
         initDisplayClass();
         initItemList();
 
-        new BukkitRunnable() {
-            @Override
-            public void run() {
-                loadShops();
-            }
-        }.runTaskLater(this.plugin, 10);
+        plugin.getFoliaLib().getScheduler().runLater(() -> {
+            loadShops();
+        }, 10);
+    }
+
+    private void disableDisplayClass() {
+        try {
+            final Class<?> clazz = Class.forName("com.snowgears.shop.display.DisplayDisabled");
+            if (AbstractDisplay.class.isAssignableFrom(clazz))
+                this.displayClass = clazz;
+        } catch (final Exception e) {
+            Shop.getPlugin().getLogger().severe("Failed to load DisplayDisabled class.");
+            Shop.getPlugin().onDisable();
+        } catch (Error e) {
+            Shop.getPlugin().getLogger().severe("Failed to load DisplayDisabled class.");
+            Shop.getPlugin().onDisable();
+        }
     }
 
     private boolean initDisplayClass(){
@@ -88,6 +119,14 @@ public class ShopHandler {
                     return true;
                 }
             } catch (final Exception e) {
+                Shop.getPlugin().getLogger().severe("Error while loading 'com.snowgears.shop.display.Display'. " + e.getMessage());
+                e.printStackTrace();
+                disableDisplayClass();
+                return false;
+            } catch (Error e) {
+                Shop.getPlugin().getLogger().severe("Error while loading 'com.snowgears.shop.display.Display'. " + e.getMessage());
+                e.printStackTrace();
+                disableDisplayClass();
                 return false;
             }
             return false;
@@ -111,8 +150,19 @@ public class ShopHandler {
                     return true;
                 }
             } catch (final Exception e) {
+                Shop.getPlugin().getLogger().severe("Error while loading com.snowgears.shop.display.Display_" + nmsVersion + " " + e.getMessage());
+                e.printStackTrace();
+                disableDisplayClass();
+                return false;
+            } catch (Error e) {
+                Shop.getPlugin().getLogger().severe("Error while loading com.snowgears.shop.display.Display_" + nmsVersion + " " + e.getMessage());
+                e.printStackTrace();
+                disableDisplayClass();
                 return false;
             }
+            
+            Shop.getPlugin().getLogger().severe("Unknown issue hooking into Minecraft Packet Classes, disabling display features.");
+            disableDisplayClass();
             return false;
         }
     }
@@ -122,7 +172,8 @@ public class ShopHandler {
             AbstractDisplay display = (AbstractDisplay) displayClass.getConstructor(Location.class).newInstance(loc);
             return display;
         } catch (Exception e){
-            e.printStackTrace();
+            // e.printStackTrace();
+            plugin.getLogger().warning("Error creating display at | World: " + loc.getWorld().getName() + " at " + loc.getX() + ", " + loc.getY() + ", " + loc.getZ());
         }
         return null;
     }
@@ -294,34 +345,26 @@ public class ShopHandler {
     public void processUnloadedShopsInChunk(Chunk chunk){
         String key = getChunkKey(chunk);
         if(unloadedShopsByChunk.containsKey(key)){
-            //System.out.println("[Shop] chunk contained unloaded shops.");
             List<UUID> playerUUIDs = new ArrayList<>();
             List<Location> shopLocations = getUnloadedShopsByChunk(key);
             for(Location shopLocation : shopLocations) {
                 AbstractShop shop = getShop(shopLocation);
                 if(shop != null){
-                    boolean signExists = shop.load();
-                    if(signExists) {
-                        //System.out.println("[Shop] shop loaded. sign exists.");
-                        //shop.getDisplay().spawn(null); //no longer spawning displays in chunk load for all players
-                        if (!playerUUIDs.contains(shop.getOwnerUUID())) {
-                            playerUUIDs.add(shop.getOwnerUUID());
+                    // Run at the shop's location to ensure it works in the correct region in Folia
+                    plugin.getFoliaLib().getScheduler().runAtLocation(shopLocation, task -> {
+                        boolean signExists = shop.load();
+                        if(signExists) {
+                            if (!playerUUIDs.contains(shop.getOwnerUUID())) {
+                                playerUUIDs.add(shop.getOwnerUUID());
+                            }
                         }
-                    }
-                    else{
-                        //System.out.println("[Shop] shop deleted. sign did not exist.");
-                        //System.out.println("[Shop] location: "+UtilMethods.getCleanLocation(shop.getSignLocation(), true));
-                        this.removeShop(shop);
-                    }
+                        else{
+                            removeShop(shop);
+                        }
+                    });
                 }
             }
             unloadedShopsByChunk.remove(key);
-
-            // No need to save all the shops here anymore
-//            //resave all shops for the player with the facing variable missing
-//            for(UUID playerUUID : playerUUIDs){
-//                saveShops(playerUUID);
-//            }
         }
     }
 
@@ -416,92 +459,356 @@ public class ShopHandler {
         return shopLocations;
     }
 
-    public HashSet<Location> getShopLocationsNearLocation(Location location){
-        int chunkX = UtilMethods.floor(location.getBlockX()) >> 4;
-        int chunkZ = UtilMethods.floor(location.getBlockZ()) >> 4;
+    /**
+     * Gets shop locations near a specific location within a default radius of 1 chunk
+     * (which covers a 3x3 chunk area)
+     * 
+     * @param location The center location to search around
+     * @return HashSet of shop locations in the surrounding chunks
+     */
+    public HashSet<Location> getShopLocationsNearLocation(Location location) {
+        return getShopLocationsNearLocation(location, plugin.getShopSearchRadius());
+    }
 
+    /**
+     * Gets shop locations near a specific location within a specified chunk radius
+     * 
+     * @param location The center location to search around
+     * @param chunkRadius The radius (in chunks) to search around the center location
+     *                    A radius of 1 means a 3x3 chunk area, 2 means 5x5, etc.
+     * @return HashSet of shop locations in the surrounding chunks
+     */
+    public HashSet<Location> getShopLocationsNearLocation(Location location, int chunkRadius) {
+        if (chunkRadius < 0) {
+            throw new IllegalArgumentException("Chunk radius cannot be negative");
+        }
+        
+        int chunkX = getChunkX(location);
+        int chunkZ = getChunkZ(location);
+        String worldName = location.getWorld().getName();
+        
         HashSet<Location> shopsNearLocation = new HashSet<>();
-        String chunkKey;
-        for(int x=-1; x<2; x++){
-            for(int z=-1; z<2; z++){
-                chunkKey = location.getWorld().getName()+"_"+(chunkX+x)+"_"+(chunkZ+z);
+        
+        // Loop through all chunks in the specified radius
+        for (int x = -chunkRadius; x <= chunkRadius; x++) {
+            for (int z = -chunkRadius; z <= chunkRadius; z++) {
+                String chunkKey = createChunkKey(worldName, chunkX + x, chunkZ + z);
                 List<Location> shopLocations = getShopLocations(chunkKey);
                 shopsNearLocation.addAll(shopLocations);
             }
         }
+        
         return shopsNearLocation;
     }
 
-//    public List<AbstractShop> getShopsNearLocation(Location location){
-//        int chunkX = UtilMethods.floor(location.getBlockX()) >> 4;
-//        int chunkZ = UtilMethods.floor(location.getBlockZ()) >> 4;
-//
-//        List<AbstractShop> shopsNearLocation = new ArrayList<>();
-//        String chunkKey;
-//        for(int x=-1; x<2; x++){
-//            for(int z=-1; z<2; z++){
-//                chunkKey = location.getWorld().getName()+"_"+(chunkX+x)+"_"+(chunkZ+z);
-//                List<Location> shopLocations = getShopLocations(chunkKey);
-//                for(Location loc : shopLocations){
-//                    shopsNearLocation.add(this.getShop(loc));
-//                }
-//            }
-//        }
-//        return shopsNearLocation;
-//    }
+    /**
+     * Gets actual shop objects near a specific location within the default radius
+     * 
+     * @param location The center location to search around
+     * @return List of shops in the surrounding chunks
+     */
+    public List<AbstractShop> getShopsNearLocation(Location location) {
+        return getShopsNearLocation(location, plugin.getShopSearchRadius());
+    }
+
+    /**
+     * Gets actual shop objects near a specific location within a specified chunk radius
+     * 
+     * @param location The center location to search around
+     * @param chunkRadius The radius (in chunks) to search around the center location
+     * @return List of shops in the surrounding chunks
+     */
+    public List<AbstractShop> getShopsNearLocation(Location location, int chunkRadius) {
+        List<AbstractShop> shopsNearLocation = new ArrayList<>();
+        
+        // Get shop locations in the specified radius
+        for (Location shopLocation : getShopLocationsNearLocation(location, chunkRadius)) {
+            AbstractShop shop = getShop(shopLocation);
+            if (shop != null) {
+                shopsNearLocation.add(shop);
+            }
+        }
+        
+        return shopsNearLocation;
+    }
+
+    /**
+     * Gets shop locations near a specific location within a specified chunk radius,
+     * filtered by maximum distance in blocks.
+     * 
+     * @param location The center location to search around
+     * @param chunkRadius The radius (in chunks) to search around the center location
+     * @param maxDistanceSquared The maximum squared distance (in blocks) to include shops
+     *                          Using squared distance avoids expensive square root calculations
+     * @return HashSet of shop locations within the distance limit
+     */
+    public HashSet<Location> getShopLocationsNearLocationWithinDistance(Location location, int chunkRadius, double maxDistanceSquared) {
+        HashSet<Location> nearbyLocations = getShopLocationsNearLocation(location, chunkRadius);
+        HashSet<Location> filteredLocations = new HashSet<>();
+        
+        // Filter by distance
+        for (Location shopLocation : nearbyLocations) {
+            // Using distanceSquared is more efficient than distance
+            if (location.distanceSquared(shopLocation) <= maxDistanceSquared) {
+                filteredLocations.add(shopLocation);
+            }
+        }
+        
+        return filteredLocations;
+    }
+
+    /**
+     * Gets actual shop objects near a specific location within a specified radius in blocks
+     * 
+     * @param location The center location to search around
+     * @param chunkRadius The radius (in chunks) to search around the center location
+     * @param maxDistance The maximum distance (in blocks) to include shops
+     * @return List of shops within the distance limit
+     */
+    public List<AbstractShop> getShopsNearLocationWithinDistance(Location location, int chunkRadius, double maxDistance) {
+        List<AbstractShop> shops = new ArrayList<>();
+        double maxDistanceSquared = maxDistance * maxDistance;
+        
+        for (Location shopLocation : getShopLocationsNearLocationWithinDistance(location, chunkRadius, maxDistanceSquared)) {
+            AbstractShop shop = getShop(shopLocation);
+            if (shop != null) {
+                shops.add(shop);
+            }
+        }
+        
+        return shops;
+    }
 
     public void processShopDisplaysNearPlayer(Player player){
-        //add the player to a logic gate for processing their displays
-        if(playersProcessingShopDisplays.contains(player.getUniqueId()))
+        // If the player is already being processed, don't start another process
+        if (playersProcessingShopDisplays.contains(player.getUniqueId())) {
             return;
+        }
+        
+        // Get current player location
+        Location currentLocation = player.getLocation();
+        
+        // Check if player has moved enough to warrant processing
+        Location lastLocation = lastProcessedLocations.get(player.getUniqueId());
+        double movementThreshold = plugin.getDisplayMovementThreshold();
+        
+        // Skip processing if player hasn't moved enough and this isn't the first check
+        if (lastLocation != null && 
+            lastLocation.getWorld().equals(currentLocation.getWorld()) && 
+            lastLocation.distanceSquared(currentLocation) < (movementThreshold * movementThreshold)) {
+            return;
+        }
+        
+        // Mark player as being processed to prevent concurrent processing
         playersProcessingShopDisplays.add(player.getUniqueId());
+        
+        // Schedule display processing task at the player's entity
+        plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+            try {
+                // Use a local variable for current location to avoid race conditions
+                Location playerLocation = player.getLocation();
+                
+                // Update the last processed location immediately to prevent multiple processings
+                lastProcessedLocations.put(player.getUniqueId(), playerLocation.clone());
+                
+                // Get all shop locations within the maximum display distance in one batch
+                HashSet<Location> nearbyShopLocations = getShopLocationsNearLocationWithinDistance(
+                    playerLocation, 
+                    plugin.getShopSearchRadius(), 
+                    plugin.getMaxShopDisplayDistance() * plugin.getMaxShopDisplayDistance()
+                );
+                
+                // Create a batch operation for all displays to minimize interference
+                // This helps prevent the "bouncing" effect when displays are created one by one
+                processBatchDisplayUpdates(player, playerLocation, nearbyShopLocations);
+                
+            } catch (Exception e) {
+                plugin.getLogger().warning("Error processing shop displays for player " + player.getName());
+                e.printStackTrace();
+            } finally {
+                // Always ensure player is removed from processing list
+                playersProcessingShopDisplays.remove(player.getUniqueId());
+            }
+        }, 1);
+    }
 
-        HashSet<Location> shopsNearPlayer = getShopLocationsNearLocation(player.getLocation());
-        if(playersWithActiveShopDisplays.containsKey(player.getUniqueId())){
-            HashSet<Location> oldShopsNearPlayer = playersWithActiveShopDisplays.get(player.getUniqueId());
-            Iterator<Location> iteratorOld = oldShopsNearPlayer.iterator();
-            while(iteratorOld.hasNext()) {
-                Location loc = iteratorOld.next();
-                if (!shopsNearPlayer.contains(loc)) {
-                    AbstractShop shop = this.getShop(loc);
-                    plugin.getServer().getScheduler().runTask(plugin, () -> {
-                        if(shop != null && player != null && player.isOnline())
-                            shop.getDisplay().remove(player);
-                    });
-                    iteratorOld.remove();
+    /**
+     * Process all shop displays in a single coordinated batch to minimize visual artifacts
+     * @param player The player to update displays for
+     * @param playerLocation The player's current location
+     * @param shopLocations Set of shop locations to process
+     */
+    private void processBatchDisplayUpdates(Player player, Location playerLocation, HashSet<Location> shopLocations) {
+        if (!player.isOnline()) return;
+        
+        // Log the processing if in debug mode
+        plugin.getLogger().debug("Processing batch display update for " + player.getName() + 
+            " at " + playerLocation.getWorld().getName() + 
+            " [" + playerLocation.getBlockX() + "," + playerLocation.getBlockY() + "," + playerLocation.getBlockZ() + "]" +
+            " with " + shopLocations.size() + " nearby shops");
+        
+        // First, collect all displays that need to be shown and those that need to be removed
+        HashSet<Location> displaysToShow = new HashSet<>();
+        HashSet<Location> displaysToRemove = new HashSet<>();
+        
+        // Determine which displays to show and which to remove
+        for (Location shopLocation : shopLocations) {
+            AbstractShop shop = getShop(shopLocation);
+            if (shop == null) continue;
+            
+            double distance = playerLocation.distance(shop.getSignLocation());
+            
+            if (distance < plugin.getMaxShopDisplayDistance()) {
+                // Within display distance, should be shown
+                displaysToShow.add(shopLocation);
+            } else {
+                // Too far, should be removed
+                displaysToRemove.add(shopLocation);
+            }
+        }
+        
+        // Also identify any current displays that are no longer in range
+        if (playersWithActiveShopDisplays.containsKey(player.getUniqueId())) {
+            HashSet<Location> activeDisplays = new HashSet<>(playersWithActiveShopDisplays.get(player.getUniqueId()));
+            for (Location displayLocation : activeDisplays) {
+                if (!shopLocations.contains(displayLocation)) {
+                    displaysToRemove.add(displayLocation);
                 }
             }
-
-            Iterator<Location> iteratorNew = shopsNearPlayer.iterator();
-            while(iteratorNew.hasNext()) {
-                Location loc = iteratorNew.next();
-                if (!oldShopsNearPlayer.contains(loc)) {
-                    AbstractShop shop = this.getShop(loc);
-                    plugin.getServer().getScheduler().runTask(plugin, () -> {
-                        if(shop != null && player != null && player.isOnline() && shop.isInitialized())
+        }
+        
+        // Process removals first to prevent interference with new spawns
+        for (Location locationToRemove : displaysToRemove) {
+            AbstractShop shop = getShop(locationToRemove);
+            if (shop != null) {
+                shop.getDisplay().remove(player);
+                removeActiveShopDisplay(player, locationToRemove);
+            }
+        }
+        
+        // Short delay before processing additions to ensure removals are complete
+        // This helps prevent the visual "refresh" effect
+        plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+            // Now process additions in priority order (closest first)
+            List<Map.Entry<Location, Double>> sortedLocations = new ArrayList<>();
+            
+            for (Location locationToShow : displaysToShow) {
+                if (!hasActiveDisplay(player, locationToShow)) {
+                    double distance = playerLocation.distance(locationToShow);
+                    sortedLocations.add(new SimpleEntry<>(locationToShow, distance));
+                }
+            }
+            
+            // Sort by distance (closest first)
+            sortedLocations.sort(Comparator.comparing(Map.Entry::getValue));
+            
+            // Process in distance order with small delays between batches to reduce visual clutter
+            // Use configurable batch size from config
+            int batchSize = plugin.getDisplayBatchSize();
+            int batchDelay = plugin.getDisplayBatchDelay();
+            int totalBatches = (sortedLocations.size() + batchSize - 1) / batchSize;
+            
+            plugin.getLogger().debug("Creating " + sortedLocations.size() + " displays in " + totalBatches + " batches for " + player.getName());
+            
+            for (int batch = 0; batch < totalBatches; batch++) {
+                final int currentBatch = batch;
+                
+                // Add a configurable delay between batches
+                plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+                    if (!player.isOnline()) return;
+                    
+                    int startIndex = currentBatch * batchSize;
+                    int endIndex = Math.min(startIndex + batchSize, sortedLocations.size());
+                    
+                    for (int i = startIndex; i < endIndex; i++) {
+                        Location locationToShow = sortedLocations.get(i).getKey();
+                        AbstractShop shop = getShop(locationToShow);
+                        
+                        if (shop != null && player.isOnline()) {
                             shop.getDisplay().spawn(player);
-                    });
-                }
+                            addActiveShopDisplay(player, locationToShow);
+                        }
+                    }
+                }, batch * batchDelay); // Configurable delay between batches
             }
-        }
-        else{
-            Iterator<Location> iteratorNew = shopsNearPlayer.iterator();
-            while(iteratorNew.hasNext()) {
-                Location loc = iteratorNew.next();
-                AbstractShop shop = this.getShop(loc);
-                plugin.getServer().getScheduler().runTaskLater(plugin, () -> {
-                    if(shop != null && player != null && player.isOnline() && shop.isInitialized())
-                        shop.getDisplay().spawn(player);
-                }, 5L); // Delay of 5 ticks to allow player to finish logging in, otherwise the item display doesn't appear!
-            }
-        }
-        playersWithActiveShopDisplays.put(player.getUniqueId(), shopsNearPlayer);
-        playersProcessingShopDisplays.remove(player.getUniqueId());
+        }, 2); // 2 tick delay after removals
     }
 
     public void clearShopDisplaysNearPlayer(Player player){
         if(playersWithActiveShopDisplays.containsKey(player.getUniqueId()))
             playersWithActiveShopDisplays.remove(player.getUniqueId());
+        
+        // Also remove player from last processed locations
+        lastProcessedLocations.remove(player.getUniqueId());
+        
+        // Also remove from processing list to avoid any potential deadlocks
+        playersProcessingShopDisplays.remove(player.getUniqueId());
+        
+        // Clear teleport cooldown as well
+        teleportCooldowns.remove(player.getUniqueId());
+
+        // Clear the cached connection when displays are removed
+        removeCachedPlayerConnection(player);
+    }
+
+    /**
+     * Force shop display processing for a player, ignoring movement threshold checks.
+     * This should be called after teleportation or world changes.
+     * 
+     * @param player The player to process shop displays for
+     */
+    public void forceProcessShopDisplaysNearPlayer(Player player) {
+        // Check if player is on teleport cooldown
+        Long lastTeleport = teleportCooldowns.get(player.getUniqueId());
+        long currentTime = System.currentTimeMillis();
+        
+        // If player is on cooldown, skip this update
+        if (lastTeleport != null && currentTime - lastTeleport < TELEPORT_COOLDOWN_MS) {
+            plugin.getLogger().debug("Skipping display update for " + player.getName() + " - on teleport cooldown");
+            return;
+        }
+        
+        // Set teleport cooldown
+        teleportCooldowns.put(player.getUniqueId(), currentTime);
+        
+        // Remove from processing list if somehow still in there
+        playersProcessingShopDisplays.remove(player.getUniqueId());
+        
+        // Remove any previous location tracking
+        lastProcessedLocations.remove(player.getUniqueId());
+        
+        plugin.getLogger().debug("Force processing shop displays for " + player.getName() + " after teleport");
+        
+        // Clear all existing displays for this player to ensure a clean slate
+        plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+            if (player.isOnline()) {
+                if (playersWithActiveShopDisplays.containsKey(player.getUniqueId())) {
+                    HashSet<Location> displays = playersWithActiveShopDisplays.get(player.getUniqueId());
+                    if (displays != null) {
+                        plugin.getLogger().debug("Removing " + displays.size() + " existing displays for " + player.getName());
+                        for (Location displayLoc : new HashSet<>(displays)) {
+                            AbstractShop shop = getShop(displayLoc);
+                            if (shop != null) {
+                                shop.getDisplay().remove(player);
+                            }
+                        }
+                    }
+                    playersWithActiveShopDisplays.remove(player.getUniqueId());
+                }
+                
+                // Process displays after a short delay to ensure all removals are complete
+                plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+                    if (player.isOnline()) {
+                        processShopDisplaysNearPlayer(player);
+                    }
+                }, 10); // Increased from 5 to 10 ticks to ensure all chunks are loaded
+            }
+        }, 1);
+    }
+
+    public boolean hasActiveDisplay(Player player, Location shopSignLocation) { 
+        HashSet<Location> shops = playersWithActiveShopDisplays.get(player.getUniqueId());
+        return shops != null && shops.contains(shopSignLocation);
     }
 
     public void addActiveShopDisplay(Player player, Location shopSignLocation){
@@ -528,19 +835,29 @@ public class ShopHandler {
         playersWithActiveShopDisplays.put(player.getUniqueId(), shops);
     }
 
-    public void addActiveShopDisplayTag(Player player, Location shopSignLocation){
-
-        if(playersActiveShopDisplayTag.containsKey(player.getUniqueId())){
+    public void addActiveShopDisplayTag(Player player, Location shopSignLocation) {
+        if (playersActiveShopDisplayTag.containsKey(player.getUniqueId())) {
             Location oldShopSignLocation = playersActiveShopDisplayTag.get(player.getUniqueId());
 
-            if(!oldShopSignLocation.equals(shopSignLocation)) {
+            if (!oldShopSignLocation.equals(shopSignLocation)) {
                 AbstractShop oldShop = getShop(oldShopSignLocation);
                 if (oldShop != null && oldShop.getDisplay() != null) {
-                    oldShop.getDisplay().removeDisplayEntities(player, true);
+                    // Use a separate task to remove the old display to avoid interference
+                    plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+                        if (player.isOnline()) {
+                            oldShop.getDisplay().removeDisplayEntities(player, true);
+                        }
+                    }, 1);
                 }
             }
         }
-        playersActiveShopDisplayTag.put(player.getUniqueId(), shopSignLocation);
+        
+        // Only update after a short delay to prevent visual glitches
+        plugin.getFoliaLib().getScheduler().runAtEntityLater(player, () -> {
+            if (player.isOnline()) {
+                playersActiveShopDisplayTag.put(player.getUniqueId(), shopSignLocation);
+            }
+        }, 2);
     }
 
 //    public boolean shopDisplayTagIsActive(Player player, Location shopSignLocation){
@@ -635,14 +952,46 @@ public class ShopHandler {
 //    }
 
     private String getChunkKey(Location location){
-        int chunkX = UtilMethods.floor(location.getBlockX()) >> 4;
-        int chunkZ = UtilMethods.floor(location.getBlockZ()) >> 4;
+        int chunkX = getChunkX(location);
+        int chunkZ = getChunkZ(location);
         String worldName = location.getWorld() != null ? location.getWorld().getName() : "unknown_world";
-        return worldName+"_"+chunkX+"_"+chunkZ;
+        return createChunkKey(worldName, chunkX, chunkZ);
     }
 
     private String getChunkKey(Chunk chunk){
-        return chunk.getWorld().getName()+"_"+chunk.getX()+"_"+chunk.getZ();
+        return createChunkKey(chunk.getWorld().getName(), chunk.getX(), chunk.getZ());
+    }
+
+    /**
+     * Creates a chunk key from world name and chunk coordinates
+     * 
+     * @param worldName The name of the world
+     * @param chunkX The x-coordinate of the chunk
+     * @param chunkZ The z-coordinate of the chunk
+     * @return A string key uniquely identifying the chunk
+     */
+    private String createChunkKey(String worldName, int chunkX, int chunkZ) {
+        return worldName + "_" + chunkX + "_" + chunkZ;
+    }
+
+    /**
+     * Gets the chunk X coordinate for a location
+     * 
+     * @param location The location
+     * @return The chunk X coordinate
+     */
+    private int getChunkX(Location location) {
+        return UtilMethods.floor(location.getBlockX()) >> 4;
+    }
+
+    /**
+     * Gets the chunk Z coordinate for a location
+     * 
+     * @param location The location
+     * @return The chunk Z coordinate
+     */
+    private int getChunkZ(Location location) {
+        return UtilMethods.floor(location.getBlockZ()) >> 4;
     }
 
     public void removeAllDisplays(Player player) {
@@ -805,7 +1154,12 @@ public class ShopHandler {
                     shopNumber++;
                 }
             }
-            plugin.getLogger().spam("    built config to save... \n" + config.saveToString());
+            
+            // Only generate the stringified config for logging if spam logging is enabled
+            if (plugin.getLogger().isLevelEnabled(ShopLogger.SPAM)) {
+                plugin.getLogger().spam("    built config to save... \n" + config.saveToString());
+            }
+            
             config.save(currentFile);
             plugin.getLogger().helpful("Saved " + shopNumber + " Shops for Player " + playerName + " to file: " + currentFile);
             return shopNumber;
@@ -859,77 +1213,74 @@ public class ShopHandler {
     }
 
     public void loadShops() {
-        plugin.getServer().getScheduler().runTaskAsynchronously(plugin, new Runnable() {
-            @Override
-            public void run() {
-                int numShopsLoaded = 0;
+        plugin.getFoliaLib().getScheduler().runAsync(task -> {
+            int numShopsLoaded = 0;
 
-                boolean convertLegacySaves = false;
-                File fileDirectory = new File(plugin.getDataFolder(), "Data");
-                if (!fileDirectory.exists())
-                    return;
+            boolean convertLegacySaves = false;
+            File fileDirectory = new File(plugin.getDataFolder(), "Data");
+            if (!fileDirectory.exists())
+                return;
 
-                // load all the yml files from the data directory
-                for (File file : fileDirectory.listFiles()) {
-                    Shop.getPlugin().getLogger().debug("Loading player shops from file: " + file.getName());
-                    try {
-                        if (file.isFile()) {
-                            if (file.getName().endsWith(".yml")
-                                    && !file.getName().contains("enderchests")
-                                    && !file.getName().contains("itemCurrency")
-                                    && !file.getName().contains("gambleDisplay")) {
-                                YamlConfiguration config = YamlConfiguration.loadConfiguration(file);
-                                boolean isLegacyConfig = false;
-                                UUID playerUUID = null;
-                                String fileNameNoExt = null;
-                                try {
-                                    int dotIndex = file.getName().lastIndexOf('.');
-                                    fileNameNoExt = file.getName().substring(0, dotIndex); //remove .yml
+            // load all the yml files from the data directory
+            for (File file : fileDirectory.listFiles()) {
+                Shop.getPlugin().getLogger().debug("Loading player shops from file: " + file.getName());
+                try {
+                    if (file.isFile()) {
+                        if (file.getName().endsWith(".yml")
+                                && !file.getName().contains("enderchests")
+                                && !file.getName().contains("itemCurrency")
+                                && !file.getName().contains("gambleDisplay")) {
+                            YamlConfiguration config = YamlConfiguration.loadConfiguration(file);
+                            boolean isLegacyConfig = false;
+                            UUID playerUUID = null;
+                            String fileNameNoExt = null;
+                            try {
+                                int dotIndex = file.getName().lastIndexOf('.');
+                                fileNameNoExt = file.getName().substring(0, dotIndex); //remove .yml
 
-                                    //all files are saved as UUID.yml except for admin shops which are admin.yml
-                                    if (!fileNameNoExt.equals("admin")) {
-                                        playerUUID = UUID.fromString(fileNameNoExt);
-                                        //file names are in UUID format. Load from new save files -> ownerUUID.yml
-                                    } else {
-                                        playerUUID = adminUUID;
-                                    }
-                                } catch (IllegalArgumentException iae) {
-                                    //file names are not in UUID format. Load from legacy save files -> ownerName + " (" + ownerUUID + ").yml
-                                    isLegacyConfig = true;
-                                    convertLegacySaves = true;
-                                    playerUUID = uidFromString(fileNameNoExt);
+                                //all files are saved as UUID.yml except for admin shops which are admin.yml
+                                if (!fileNameNoExt.equals("admin")) {
+                                    playerUUID = UUID.fromString(fileNameNoExt);
+                                    //file names are in UUID format. Load from new save files -> ownerUUID.yml
+                                } else {
+                                    playerUUID = adminUUID;
                                 }
-                                numShopsLoaded += loadShopsFromConfig(config, isLegacyConfig);
-                                if (isLegacyConfig) {
-                                    //save new file
-                                    saveShops(playerUUID, true);
-                                    //delete old file
-                                    file.delete();
-                                }
-                                if (plugin.getDebug_forceResaveAll()) {
-                                    saveShops(playerUUID, true);
-                                }
+                            } catch (IllegalArgumentException iae) {
+                                //file names are not in UUID format. Load from legacy save files -> ownerName + " (" + ownerUUID + ").yml
+                                isLegacyConfig = true;
+                                convertLegacySaves = true;
+                                playerUUID = uidFromString(fileNameNoExt);
+                            }
+                            numShopsLoaded += loadShopsFromConfig(config, isLegacyConfig);
+                            if (isLegacyConfig) {
+                                //save new file
+                                saveShops(playerUUID, true);
+                                //delete old file
+                                file.delete();
+                            }
+                            if (plugin.getDebug_forceResaveAll()) {
+                                saveShops(playerUUID, true);
                             }
                         }
                     }
-                    catch (Exception e) {
-                        Shop.getPlugin().getLogger().log(Level.SEVERE, "Error while loading Shop from player file (" + file.getName() + "), please report this error to the Shop developers: " + e.getMessage());
-                        e.printStackTrace();
-                    }
                 }
-                if(convertLegacySaves)
-                    convertLegacyShopSaves();
-
-                Shop.getPlugin().getLogger().log(Level.INFO, "Loaded " + numShopsLoaded + " Shops!");
-
-                //dont refresh displays at load time anymore. they are now loaded in client side on login
-//                new BukkitRunnable() {
-//                    @Override
-//                    public void run() {
-//                        refreshShopDisplays(null);
-//                    }
-//                }.runTaskLater(plugin, 20);
+                catch (Exception e) {
+                    Shop.getPlugin().getLogger().log(Level.SEVERE, "Error while loading Shop from player file (" + file.getName() + "), please report this error to the Shop developers: " + e.getMessage());
+                    e.printStackTrace();
+                }
             }
+            if(convertLegacySaves)
+                convertLegacyShopSaves();
+
+            Shop.getPlugin().getLogger().log(Level.INFO, "Loaded " + numShopsLoaded + " Shops!");
+
+            //dont refresh displays at load time anymore. they are now loaded in client side on login
+    //                new BukkitRunnable() {
+    //                    @Override
+    //                    public void run() {
+    //                        refreshShopDisplays(null);
+    //                    }
+    //                }.runTaskLater(plugin, 20);
         });
     }
 
@@ -1022,13 +1373,10 @@ public class ShopHandler {
                         //if chunk its in is already loaded, calculate it here
                         if(shop.getDisplay().isChunkLoaded()) {
                             //run this task synchronously
-                            Bukkit.getScheduler().runTask(plugin, new Runnable() {
-                                @Override
-                                public void run() {
-                                    boolean signDefined = shop.load();
-                                    if(signDefined)
-                                        addShop(shop);
-                                }
+                            plugin.getFoliaLib().getScheduler().runAtLocation(shop.getSignLocation(), task -> {
+                                boolean signDefined = shop.load();
+                                if(signDefined)
+                                    addShop(shop);
                             });
                         }
                         //if the chunk is not already loaded, add it to a list to calculate it at chunkloadevent later
@@ -1198,5 +1546,123 @@ public class ShopHandler {
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+
+    /**
+     * Checks if a player is within the specified distance of a chunk
+     * 
+     * @param player The player to check
+     * @param chunk The chunk to check against
+     * @param maxDistance The maximum distance in blocks
+     * @return True if the player is near the chunk
+     */
+    private boolean isPlayerNearChunk(Player player, Chunk chunk, double maxDistance) {
+        if (!player.getWorld().equals(chunk.getWorld())) {
+            return false;
+        }
+        
+        // Get chunk boundaries
+        int minX = chunk.getX() << 4;
+        int minZ = chunk.getZ() << 4;
+        int maxX = minX + 15;
+        int maxZ = minZ + 15;
+        
+        // Get player location
+        Location playerLoc = player.getLocation();
+        int playerX = playerLoc.getBlockX();
+        int playerZ = playerLoc.getBlockZ();
+        
+        // Calculate minimum distance to chunk border
+        double distance;
+        
+        // Player is outside chunk on X axis
+        if (playerX < minX) {
+            distance = minX - playerX;
+        } else if (playerX > maxX) {
+            distance = playerX - maxX;
+        } else {
+            // Player is within chunk X range
+            distance = 0;
+        }
+        
+        // Player is outside chunk on Z axis
+        if (playerZ < minZ) {
+            double zDistance = minZ - playerZ;
+            distance = Math.sqrt(distance * distance + zDistance * zDistance);
+        } else if (playerZ > maxZ) {
+            double zDistance = playerZ - maxZ;
+            distance = Math.sqrt(distance * distance + zDistance * zDistance);
+        }
+        
+        return distance <= maxDistance;
+    }
+    
+    /**
+     * Rebuilds all shop displays in a chunk for nearby players
+     * Called after a chunk has been loaded to ensure displays are shown
+     * 
+     * @param chunk The chunk that was loaded
+     */
+    public void rebuildDisplaysInChunk(Chunk chunk) {
+        // Get shop locations in this chunk
+        String chunkKey = getChunkKey(chunk);
+        List<Location> shopLocations = getShopLocations(chunkKey);
+        
+        // Only proceed if this chunk has shops
+        if (shopLocations.isEmpty()) {
+            return;
+        }
+        
+        // Process for all players who might be able to see shops in this chunk
+        for (Player player : chunk.getWorld().getPlayers()) {
+            // Skip players who recently teleported
+            Long lastTeleport = teleportCooldowns.get(player.getUniqueId());
+            if (lastTeleport != null && System.currentTimeMillis() - lastTeleport < TELEPORT_COOLDOWN_MS) {
+                plugin.getLogger().debug("Skipping chunk display update for " + player.getName() + " - on teleport cooldown");
+                continue;
+            }
+            
+            if (isPlayerNearChunk(player, chunk, plugin.getMaxShopDisplayDistance())) {
+                plugin.getLogger().debug("Rebuilding shop displays for " + player.getName() + " in chunk " + chunkKey);
+                
+                // Don't force a refresh - just run the normal process which respects all the checks
+                processShopDisplaysNearPlayer(player);
+            }
+        }
+    }
+
+    /**
+     * Gets the cached player connection for packet sending
+     * @param player The player to get connection for
+     * @return The player's network connection object
+     */
+    public Object getCachedPlayerConnection(Player player) {
+        UUID playerId = player.getUniqueId();
+        // Check if we have a cached connection
+        Object connection = playerConnectionCache.get(playerId);
+        if (connection == null) {
+            // If not cached, get it and store it
+            connection = plugin.getNmsBullshitHandler().getPlayerConnection(player);
+            if (connection != null) {
+                playerConnectionCache.put(playerId, connection);
+            }
+        }
+        return connection;
+    }
+
+    /**
+     * Removes the cached player connection
+     * @param player The player whose connection to remove
+     */
+    public void removeCachedPlayerConnection(Player player) {
+        playerConnectionCache.remove(player.getUniqueId());
+    }
+
+    /**
+     * Removes the cached player connection by UUID
+     * @param playerId UUID of the player whose connection to remove
+     */
+    public void removeCachedPlayerConnection(UUID playerId) {
+        playerConnectionCache.remove(playerId);
     }
 }
