@@ -166,24 +166,22 @@ public class MiscListener implements Listener {
                 }
 
                 //give player a limited amount of time to finish creating the shop until it is deleted
-                plugin.getServer().getScheduler().scheduleSyncDelayedTask(plugin, new Runnable() {
-                    public void run() {
-                        //the shop has still not been initialized with an item from a player
-                        if (!shop.isInitialized()) {
-                            plugin.getShopHandler().removeShop(shop);
-                            if (b.getBlockData() instanceof WallSign) {
-                                String[] lines = ShopMessage.getTimeoutSignLines(shop);
-                                Sign sign = (Sign) b.getState();
-                                sign.setLine(0, lines[0]);
-                                sign.setLine(1, lines[1]);
-                                sign.setLine(2, lines[2]);
-                                sign.setLine(3, lines[3]);
-                                sign.update(true);
-                                cancelShopCreationProcess(player);
-                            }
+                plugin.getFoliaLib().getScheduler().runLater(() -> {
+                    //the shop has still not been initialized with an item from a player
+                    if (!shop.isInitialized()) {
+                        plugin.getShopHandler().removeShop(shop);
+                        if (b.getBlockData() instanceof WallSign) {
+                            String[] lines = ShopMessage.getTimeoutSignLines(shop);
+                            Sign sign = (Sign) b.getState();
+                            sign.setLine(0, lines[0]);
+                            sign.setLine(1, lines[1]);
+                            sign.setLine(2, lines[2]);
+                            sign.setLine(3, lines[3]);
+                            sign.update(true);
+                            cancelShopCreationProcess(player);
                         }
                     }
-                }, 30 * 20); //30 seconds * 20 ticks
+                }, 30 * 20); // 30 seconds * 20 ticks
             }
         }
     }
@@ -292,8 +290,10 @@ public class MiscListener implements Listener {
                                 return;
                             }
 
-                            if(!plugin.getShopCreationUtil().shopCanBeCreated(player, clicked))
+                            if(!plugin.getShopCreationUtil().shopCanBeCreated(player, clicked)){
+                                event.setCancelled(true);
                                 return;
+                            }
                             BlockFace signFacing = plugin.getShopCreationUtil().calculateBlockFaceForSign(player, clicked, event.getBlockFace());
                             if(signFacing == null) {
                                 event.setCancelled(true);
@@ -353,8 +353,10 @@ public class MiscListener implements Listener {
                     return;
                 }
 
-                if(!plugin.getShopCreationUtil().shopCanBeCreated(player, clicked))
+                if(!plugin.getShopCreationUtil().shopCanBeCreated(player, clicked)){
+                    event.setCancelled(true);
                     return;
+                }
 
                 event.setCancelled(true);
                 BlockFace signFacing = plugin.getShopCreationUtil().calculateBlockFaceForSign(player, clicked, event.getBlockFace());
@@ -372,23 +374,23 @@ public class MiscListener implements Listener {
                 process.displayFloatingText("createHitChest", null);
                 List<String> autocomplete = new ArrayList<>();
                 Arrays.asList(ShopType.values()).forEach((shopType -> autocomplete.add(shopType.toString().toLowerCase())));
-                player.setCustomChatCompletions(autocomplete);
+                try {
+                    player.setCustomChatCompletions(autocomplete);
+                } catch (Error error) {} // Suppress error if autocomplete is not supported
                 if((!plugin.usePerms() && player.isOp()) || (plugin.usePerms() && player.hasPermission("shop.operator"))) {
                     ShopMessage.sendMessage("adminCreateHitChest", null, process, player);
                 }
 
                 //give player a limited amount of time to finish creating the shop until it is deleted
                 final UUID originalProcessUUID = process.getUniqueID();
-                plugin.getServer().getScheduler().scheduleSyncDelayedTask(plugin, new Runnable() {
-                    public void run() {
-                        //the shop has still not been initialized with an item from a player
-                        ShopCreationProcess process = playerChatCreationSteps.get(player.getUniqueId());
-                        if (process != null && process.getUniqueID().equals(originalProcessUUID)) {
-                            process.cleanup();
-                            playerChatCreationSteps.remove(player.getUniqueId());
-                            plugin.getCreativeSelectionListener().removePlayerFromCreativeSelection(player);
-                            ShopMessage.sendMessage("interactionIssue", "createHitChestTimeout", process, player);
-                        }
+                plugin.getFoliaLib().getScheduler().runLater(() -> {
+                    //the shop has still not been initialized with an item from a player
+                    ShopCreationProcess currentProcess = playerChatCreationSteps.get(player.getUniqueId());
+                    if (currentProcess != null && currentProcess.getUniqueID().equals(originalProcessUUID)) {
+                        currentProcess.cleanup();
+                        playerChatCreationSteps.remove(player.getUniqueId());
+                        plugin.getCreativeSelectionListener().removePlayerFromCreativeSelection(player);
+                        ShopMessage.sendMessage("interactionIssue", "createHitChestTimeout", currentProcess, player);
                     }
                 }, 30 * 20); // 30 seconds * 20 ticks
             }
@@ -400,6 +402,7 @@ public class MiscListener implements Listener {
         Player player = event.getPlayer();
         if(playerChatCreationSteps.containsKey(player.getUniqueId())){
             ShopCreationProcess process = playerChatCreationSteps.get(player.getUniqueId());
+            plugin.getLogger().debug("Shop Creation Process: " + process.getStep() + " Player " + player.getName() + " input: " + event.getMessage(), true);
             switch (process.getStep()){
                 case SHOP_TYPE:
                     ShopType type = plugin.getShopCreationUtil().getShopType(event.getMessage());
