@@ -12,7 +12,7 @@ import org.bukkit.OfflinePlayer;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
-
+import org.bukkit.ChatColor;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -142,46 +142,43 @@ public class LogHandler {
             plugin.getLogger().notice(
                     player.getName() + " created a " + shop.getType().name().toUpperCase() + " shop at ("
                             + "x: " + shop.getChestLocation().getBlockX() + " y: " + shop.getChestLocation().getBlockY() + " z: " + shop.getChestLocation().getBlockZ()
-                            + ") item: " + new ItemNameUtil().getName(shop.getItemStack()).toPlainText()
-                            + (shop.getSecondaryItemStack() != null ? " barterItem: " + new ItemNameUtil().getName(shop.getSecondaryItemStack()).toPlainText() : "")
+                            + ") item: " + ChatColor.stripColor(new ItemNameUtil().getName(shop.getItemStack()).toPlainText())
+                            + (shop.getSecondaryItemStack() != null ? " barterItem: " + ChatColor.stripColor(new ItemNameUtil().getName(shop.getSecondaryItemStack()).toPlainText()) : "")
             );
         }
         if (actionType == ShopActionType.DESTROY) {
             plugin.getLogger().notice(
                     player.getName() + " destroyed a " + shop.getType().name().toUpperCase() + " shop at ("
                             + "x: " + shop.getChestLocation().getBlockX() + " y: " + shop.getChestLocation().getBlockY() + " z: " + shop.getChestLocation().getBlockZ()
-                            + ") item: " + new ItemNameUtil().getName(shop.getItemStack()).toPlainText()
-                            + (shop.getSecondaryItemStack() != null ? " barterItem: " + new ItemNameUtil().getName(shop.getSecondaryItemStack()).toPlainText() : "")
+                            + ") item: " + ChatColor.stripColor(new ItemNameUtil().getName(shop.getItemStack()).toPlainText())
+                            + (shop.getSecondaryItemStack() != null ? " barterItem: " + ChatColor.stripColor(new ItemNameUtil().getName(shop.getSecondaryItemStack()).toPlainText()) : "")
             );
         }
 
         if (!this.enabled) return;
-        Bukkit.getScheduler().runTaskAsynchronously(plugin, new Runnable() {
-            @Override
-            public void run() {
-                // Connect to datasource & create statement in "try" to handle automatically closing the connection!
-                try (Connection conn = dataSource.getConnection();
-                    PreparedStatement stmt = conn.prepareStatement("INSERT INTO shop_action(ts, player_uuid, owner_uuid, shop_uuid, player_action, shop_world, shop_x, shop_y, shop_z) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?);");
-                ){
-                    stmt.setTimestamp(1, new Timestamp(new Date().getTime()));
-                    stmt.setString(2, player.getUniqueId().toString());
-                    if (shop.getOwnerUUID().equals(plugin.getShopHandler().getAdminUUID()))
-                        stmt.setString(3, "admin");
-                    else
-                        stmt.setString(3, shop.getOwnerUUID().toString());
-                    String shop_uuid = "";
-                    if (shop.getId() != null && shop.getId().toString() != null) shop_uuid = shop.getId().toString();
-                    stmt.setString(4, shop_uuid);
-                    stmt.setString(5, actionType.toString());
-                    stmt.setString(6, shop.getSignLocation().getWorld().getName());
-                    stmt.setInt(7, shop.getSignLocation().getBlockX());
-                    stmt.setInt(8, shop.getSignLocation().getBlockY());
-                    stmt.setInt(9, shop.getSignLocation().getBlockZ());
-                    stmt.execute();
-                } catch (SQLException e) {
-                    plugin.getLogger().log(Level.WARNING, "SQL error occurred while trying to log player action.");
-                    e.printStackTrace();
-                }
+        plugin.getFoliaLib().getScheduler().runAsync(task -> {
+            // Connect to datasource & create statement in "try" to handle automatically closing the connection!
+            try (Connection conn = dataSource.getConnection();
+                PreparedStatement stmt = conn.prepareStatement("INSERT INTO shop_action(ts, player_uuid, owner_uuid, shop_uuid, player_action, shop_world, shop_x, shop_y, shop_z) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?);");
+            ){
+                stmt.setTimestamp(1, new Timestamp(new Date().getTime()));
+                stmt.setString(2, player.getUniqueId().toString());
+                if (shop.getOwnerUUID().equals(plugin.getShopHandler().getAdminUUID()))
+                    stmt.setString(3, "admin");
+                else
+                    stmt.setString(3, shop.getOwnerUUID().toString());
+                String shop_uuid = "";
+                if (shop.getId() != null && shop.getId().toString() != null) shop_uuid = shop.getId().toString();
+                stmt.setString(4, shop_uuid);
+                stmt.setString(5, actionType.toString());
+                stmt.setString(6, shop.getSignLocation().getWorld().getName());
+                stmt.setInt(7, shop.getSignLocation().getBlockX());
+                stmt.setInt(8, shop.getSignLocation().getBlockY());
+                stmt.setInt(9, shop.getSignLocation().getBlockZ());
+                stmt.execute();
+            } catch (SQLException e) {
+                plugin.getLogger().log(Level.WARNING, "SQL error occurred while trying to log player action.");
+                e.printStackTrace();
             }
         });
     }
@@ -189,8 +186,8 @@ public class LogHandler {
     public void logTransaction(Player player, AbstractShop shop, ShopType transactionType, double price, int amount){
         plugin.getLogger().helpful(
             "Shop " + shop.getType().name().toUpperCase() + " from/to " + player.getName() + ": "
-                + new ItemNameUtil().getName(shop.getItemStack()).toPlainText() + "(x" + amount + ")" + " for " + plugin.getPriceString(price, true)
-                + " | Shop owned by " + plugin.getServer().getOfflinePlayer(shop.getOwnerUUID()).getName() + " at (x: " + shop.getChestLocation().getBlockX() + " y: " + shop.getChestLocation().getBlockY() + " z: " + shop.getChestLocation().getBlockZ() + ")"
+                + ChatColor.stripColor(new ItemNameUtil().getName(shop.getItemStack()).toPlainText()) + "(x" + amount + ")" + " for " + plugin.getPriceString(price, true)
+                + " | Shop owned by " + shop.getOwnerName() + " at (x: " + shop.getChestLocation().getBlockX() + " y: " + shop.getChestLocation().getBlockY() + " z: " + shop.getChestLocation().getBlockZ() + ")"
         );
 
         try {
@@ -199,64 +196,61 @@ public class LogHandler {
         } catch (Exception e) { /* Ignore errors, added for safety. */ }
 
         if(!enabled) return;
-        Bukkit.getScheduler().runTaskAsynchronously(plugin, new Runnable() {
-            @Override
-            public void run() {
-                // Log the Transaction that occured
-                int transactionID = 0;
-                // Connect to datasource & create statement in "try" to handle automatically closing the connection!
-                try (
-                    Connection conn = dataSource.getConnection();
-                    PreparedStatement logTxStmt = conn.prepareStatement("INSERT INTO shop_transaction (t_type, price, amount, item, barter_item) VALUES(?, ?, ?, ?, ?);", Statement.RETURN_GENERATED_KEYS);
-                ) {
-                    logTxStmt.setString(1, transactionType.toString().toUpperCase());
-                    logTxStmt.setDouble(2, price);
-                    logTxStmt.setInt(3, amount);
-                    logTxStmt.setString(4, UtilMethods.itemStackToBase64(shop.getItemStack()));
-                    if (shop.getSecondaryItemStack() != null)
-                        logTxStmt.setString(5, UtilMethods.itemStackToBase64(shop.getSecondaryItemStack()));
-                    else
-                        logTxStmt.setNull(5, Types.VARCHAR);
+        plugin.getFoliaLib().getScheduler().runAsync(task -> {
+            // Log the Transaction that occured
+            int transactionID = 0;
+            // Connect to datasource & create statement in "try" to handle automatically closing the connection!
+            try (
+                Connection conn = dataSource.getConnection();
+                PreparedStatement logTxStmt = conn.prepareStatement("INSERT INTO shop_transaction (t_type, price, amount, item, barter_item) VALUES(?, ?, ?, ?, ?);", Statement.RETURN_GENERATED_KEYS);
+            ) {
+                logTxStmt.setString(1, transactionType.toString().toUpperCase());
+                logTxStmt.setDouble(2, price);
+                logTxStmt.setInt(3, amount);
+                logTxStmt.setString(4, UtilMethods.itemStackToBase64(shop.getItemStack()));
+                if (shop.getSecondaryItemStack() != null)
+                    logTxStmt.setString(5, UtilMethods.itemStackToBase64(shop.getSecondaryItemStack()));
+                else
+                    logTxStmt.setNull(5, Types.VARCHAR);
 
-                    logTxStmt.execute();
-                    ResultSet txRS = logTxStmt.getGeneratedKeys();
-                    txRS.next();
-                    transactionID = txRS.getInt(1);
-                } catch (SQLException e) {
-                    plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to log transaction/shop action.");
-                    e.printStackTrace();
-                } catch (IOException e) {
-                    plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to log transaction/shop action. Issue with converting itemstack to base64!");
-                    e.printStackTrace();
-                }
+                logTxStmt.execute();
+                ResultSet txRS = logTxStmt.getGeneratedKeys();
+                txRS.next();
+                transactionID = txRS.getInt(1);
+            } catch (SQLException e) {
+                plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to log transaction/shop action.");
+                e.printStackTrace();
+            } catch (IOException e) {
+                plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to log transaction/shop action. Issue with converting itemstack to base64!");
+                e.printStackTrace();
+            }
 
-                // Log the action that occured
-                if (transactionID == 0) return; // Last query failed, so skip this one!
-                // Connect to datasource & create statement in "try" to handle automatically closing the connection!
-                try (
-                    Connection conn = dataSource.getConnection();
-                    PreparedStatement actionStmt = conn.prepareStatement("INSERT INTO shop_action(ts, player_uuid, owner_uuid, shop_uuid, player_action, transaction_id, shop_world, shop_x, shop_y, shop_z) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?);");
-                ) {
-                    actionStmt.setTimestamp(1, new Timestamp(new Date().getTime()));
-                    actionStmt.setString(2, player.getUniqueId().toString());
-                    if(shop.getOwnerUUID().equals(plugin.getShopHandler().getAdminUUID()))
-                        actionStmt.setString(3, "admin");
-                    else
-                        actionStmt.setString(3, shop.getOwnerUUID().toString());
-                    String shop_uuid = "";
-                    if (shop.getId() != null && shop.getId().toString() != null) shop_uuid = shop.getId().toString();
-                    actionStmt.setString(4, shop_uuid);
-                    actionStmt.setString(5, ShopActionType.TRANSACT.toString());
-                    actionStmt.setInt(6, transactionID);
-                    actionStmt.setString(7, shop.getSignLocation().getWorld().getName());
-                    actionStmt.setInt(8, shop.getSignLocation().getBlockX());
-                    actionStmt.setInt(9, shop.getSignLocation().getBlockY());
-                    actionStmt.setInt(10, shop.getSignLocation().getBlockZ());
-                    actionStmt.execute();
-                } catch (SQLException e) {
-                    plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to log transaction/shop action.");
-                    e.printStackTrace();
-                }
+            // Log the action that occured
+            if (transactionID == 0) return; // Last query failed, so skip this one!
+            // Connect to datasource & create statement in "try" to handle automatically closing the connection!
+            try (
+                Connection conn = dataSource.getConnection();
+                PreparedStatement actionStmt = conn.prepareStatement("INSERT INTO shop_action(ts, player_uuid, owner_uuid, shop_uuid, player_action, transaction_id, shop_world, shop_x, shop_y, shop_z) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?);");
+            ) {
+                actionStmt.setTimestamp(1, new Timestamp(new Date().getTime()));
+                actionStmt.setString(2, player.getUniqueId().toString());
+                if(shop.getOwnerUUID().equals(plugin.getShopHandler().getAdminUUID()))
+                    actionStmt.setString(3, "admin");
+                else
+                    actionStmt.setString(3, shop.getOwnerUUID().toString());
+                String shop_uuid = "";
+                if (shop.getId() != null && shop.getId().toString() != null) shop_uuid = shop.getId().toString();
+                actionStmt.setString(4, shop_uuid);
+                actionStmt.setString(5, ShopActionType.TRANSACT.toString());
+                actionStmt.setInt(6, transactionID);
+                actionStmt.setString(7, shop.getSignLocation().getWorld().getName());
+                actionStmt.setInt(8, shop.getSignLocation().getBlockX());
+                actionStmt.setInt(9, shop.getSignLocation().getBlockY());
+                actionStmt.setInt(10, shop.getSignLocation().getBlockZ());
+                actionStmt.execute();
+            } catch (SQLException e) {
+                plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to log transaction/shop action.");
+                e.printStackTrace();
             }
         });
     }
@@ -268,106 +262,103 @@ public class LogHandler {
         }
         offlineTransactions.setIsCalculating(true);
 
-        Bukkit.getScheduler().runTaskAsynchronously(plugin, new Runnable() {
-            @Override
-            public void run() {
-                String query = "SELECT * from shop_action RIGHT JOIN shop_transaction on shop_action.transaction_id = shop_transaction.id where owner_uuid=? and ts > ?;";
-                // Connect to datasource & create statement in "try" to handle automatically closing the connection!
-                try (
-                    Connection conn = dataSource.getConnection();
-                    PreparedStatement stmt = conn.prepareStatement(query);
-                ) {
-                    // Initialize variables to accumulate totals
-                    double totalProfit = 0.0;
-                    double totalSpent = 0.0;
-                    Map<ItemStack, Integer> itemsBought = new HashMap<>();
-                    Map<ItemStack, Integer> itemsSold = new HashMap<>();
+        plugin.getFoliaLib().getScheduler().runAsync(task -> {
+            String query = "SELECT * from shop_action RIGHT JOIN shop_transaction on shop_action.transaction_id = shop_transaction.id where owner_uuid=? and ts > ?;";
+            // Connect to datasource & create statement in "try" to handle automatically closing the connection!
+            try (
+                Connection conn = dataSource.getConnection();
+                PreparedStatement stmt = conn.prepareStatement(query);
+            ) {
+                // Initialize variables to accumulate totals
+                double totalProfit = 0.0;
+                double totalSpent = 0.0;
+                Map<ItemStack, Integer> itemsBought = new HashMap<>();
+                Map<ItemStack, Integer> itemsSold = new HashMap<>();
 
-                    // Prepare and execute the statement
+                // Prepare and execute the statement
 
-                    stmt.setString(1, offlineTransactions.getPlayerUUID().toString());
-                    stmt.setTimestamp(2, new Timestamp(offlineTransactions.getLastPlayed()));
-                    ResultSet resultSet = stmt.executeQuery();
+                stmt.setString(1, offlineTransactions.getPlayerUUID().toString());
+                stmt.setTimestamp(2, new Timestamp(offlineTransactions.getLastPlayed()));
+                ResultSet resultSet = stmt.executeQuery();
 
-                    int size = 0;
-                    if (resultSet != null) {
-                        while (resultSet.next()) {
-                            size++;
+                int size = 0;
+                if (resultSet != null) {
+                    while (resultSet.next()) {
+                        size++;
 
-                            // Extract transaction data
-                            String purchaserUUID = resultSet.getString("player_uuid");
-                            String tType = resultSet.getString("t_type");
-                            double price = resultSet.getDouble("price");
-                            int amount = resultSet.getInt("amount");
-                            String item = resultSet.getString("item");
-                            String barterItem = resultSet.getString("barter_item"); // May be null
+                        // Extract transaction data
+                        String purchaserUUID = resultSet.getString("player_uuid");
+                        String tType = resultSet.getString("t_type");
+                        double price = resultSet.getDouble("price");
+                        int amount = resultSet.getInt("amount");
+                        String item = resultSet.getString("item");
+                        String barterItem = resultSet.getString("barter_item"); // May be null
 
-                            String shopWorld = resultSet.getString("shop_world");
-                            int shopX = resultSet.getInt("shop_x");
-                            int shopY = resultSet.getInt("shop_y");
-                            int shopZ = resultSet.getInt("shop_z");
+                        String shopWorld = resultSet.getString("shop_world");
+                        int shopX = resultSet.getInt("shop_x");
+                        int shopY = resultSet.getInt("shop_y");
+                        int shopZ = resultSet.getInt("shop_z");
 
-                            Location loc = new Location(Bukkit.getWorld(shopWorld), shopX, shopY, shopZ);
+                        Location loc = new Location(Bukkit.getWorld(shopWorld), shopX, shopY, shopZ);
 
-                            ItemStack itemstack = UtilMethods.itemStackFromBase64(item);
-                            ItemStack barterItemstack = null;
+                        ItemStack itemstack = UtilMethods.itemStackFromBase64(item);
+                        ItemStack barterItemstack = null;
 
-                            // Create item stack to use for the Maps (show all identical items as the same sold row)
-                            ItemStack itemCheckClone = itemstack.clone();
-                            itemCheckClone.setAmount(1);
+                        // Create item stack to use for the Maps (show all identical items as the same sold row)
+                        ItemStack itemCheckClone = itemstack.clone();
+                        itemCheckClone.setAmount(1);
 
-                            // Process transactions based on their type
-                            if (tType.equalsIgnoreCase("BUY")) {
-                                // User spent money to buy items
-                                totalSpent += price;
-                                int itemsBoughtAmt = itemsBought.getOrDefault(itemCheckClone, 0) + amount;
-                                itemsBought.put(itemCheckClone,itemsBoughtAmt);
-                            } else if (tType.equalsIgnoreCase("SELL")) {
-                                // User earned money by selling items
-                                totalProfit += price;
-                                int itemsSoldAmt = itemsSold.getOrDefault(itemCheckClone, 0) + amount;
-                                itemsSold.put(itemCheckClone, itemsSoldAmt);
-                            } else if (tType.equalsIgnoreCase("BARTER")) {
-                                int itemsAmt = itemsSold.getOrDefault(itemCheckClone, 0) + amount;
-                                itemsSold.put(itemCheckClone, itemsAmt);
-                                if (barterItem != null) {
-                                    // Barter Item is the "currency" item being spent in the transaction
-                                    barterItemstack = UtilMethods.itemStackFromBase64(barterItem);
-                                    ItemStack barterItemCheckClone = barterItemstack.clone();
-                                    barterItemCheckClone.setAmount(1);
-                                    itemsBought.put(barterItemCheckClone, itemsBought.getOrDefault(barterItemCheckClone, 0) + (int) price);
-                                }
+                        // Process transactions based on their type
+                        if (tType.equalsIgnoreCase("BUY")) {
+                            // User spent money to buy items
+                            totalSpent += price;
+                            int itemsBoughtAmt = itemsBought.getOrDefault(itemCheckClone, 0) + amount;
+                            itemsBought.put(itemCheckClone,itemsBoughtAmt);
+                        } else if (tType.equalsIgnoreCase("SELL")) {
+                            // User earned money by selling items
+                            totalProfit += price;
+                            int itemsSoldAmt = itemsSold.getOrDefault(itemCheckClone, 0) + amount;
+                            itemsSold.put(itemCheckClone, itemsSoldAmt);
+                        } else if (tType.equalsIgnoreCase("BARTER")) {
+                            int itemsAmt = itemsSold.getOrDefault(itemCheckClone, 0) + amount;
+                            itemsSold.put(itemCheckClone, itemsAmt);
+                            if (barterItem != null) {
+                                // Barter Item is the "currency" item being spent in the transaction
+                                barterItemstack = UtilMethods.itemStackFromBase64(barterItem);
+                                ItemStack barterItemCheckClone = barterItemstack.clone();
+                                barterItemCheckClone.setAmount(1);
+                                itemsBought.put(barterItemCheckClone, itemsBought.getOrDefault(barterItemCheckClone, 0) + (int) price);
                             }
-
-                            OfflinePlayer purchaser = null;
-                            if (purchaserUUID != null && !purchaserUUID.isEmpty()) {
-                                purchaser = Bukkit.getServer().getOfflinePlayer(UUID.fromString(purchaserUUID));
-                            }
-
-                            offlineTransactions.addTx(loc, ShopType.valueOf(tType), price, purchaser, amount, itemstack, barterItemstack);
                         }
-                    }
 
-                    // Update offlineTransactions object with the calculated data
-                    offlineTransactions.setNumTransactions(size);
-                    offlineTransactions.setTotalProfit(totalProfit);
-                    offlineTransactions.setTotalSpent(totalSpent);
-                    offlineTransactions.setItemsBought(itemsBought);
-                    offlineTransactions.setItemsSold(itemsSold);
-                    offlineTransactions.setIsCalculating(false);
-                } catch (SQLException e){
-                    plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to get offline transactions.");
-                    e.printStackTrace();
-                    offlineTransactions.setIsCalculating(false);
-                } catch (IOException e) {
-                    plugin.getLogger().log(Level.WARNING,"IOException occurred while trying to get offline transactions. Unable to parse itemstack from base64!");
-                    e.printStackTrace();
-                    offlineTransactions.setIsCalculating(false);
-                } catch (ClassNotFoundException e) {
-                    plugin.getLogger().log(Level.WARNING,"ClassNotFoundException occurred while trying to get offline transactions. Unable to parse itemstack from base64!");
-                    e.printStackTrace();
-                    offlineTransactions.setIsCalculating(false);
+                        OfflinePlayer purchaser = null;
+                        if (purchaserUUID != null && !purchaserUUID.isEmpty()) {
+                            purchaser = Bukkit.getServer().getOfflinePlayer(UUID.fromString(purchaserUUID));
+                        }
+
+                        offlineTransactions.addTx(loc, ShopType.valueOf(tType), price, purchaser, amount, itemstack, barterItemstack);
+                    }
                 }
+
+                // Update offlineTransactions object with the calculated data
+                offlineTransactions.setNumTransactions(size);
+                offlineTransactions.setTotalProfit(totalProfit);
+                offlineTransactions.setTotalSpent(totalSpent);
+                offlineTransactions.setItemsBought(itemsBought);
+                offlineTransactions.setItemsSold(itemsSold);
+                offlineTransactions.setIsCalculating(false);
+            } catch (SQLException e){
+                plugin.getLogger().log(Level.WARNING,"SQL error occurred while trying to get offline transactions.");
+                e.printStackTrace();
+                offlineTransactions.setIsCalculating(false);
+            } catch (IOException e) {
+                plugin.getLogger().log(Level.WARNING,"IOException occurred while trying to get offline transactions. Unable to parse itemstack from base64!");
+                e.printStackTrace();
+                offlineTransactions.setIsCalculating(false);
+            } catch (ClassNotFoundException e) {
+                plugin.getLogger().log(Level.WARNING,"ClassNotFoundException occurred while trying to get offline transactions. Unable to parse itemstack from base64!");
+                e.printStackTrace();
+                offlineTransactions.setIsCalculating(false);
             }
         });
     }
