@@ -232,7 +232,7 @@ public class ShopCreationUtil {
     }
 
     public void sendCreationSuccess(Player player, AbstractShop shop){
-        shop.getDisplay().spawn(player);
+        if (shop.getDisplay() != null) shop.getDisplay().spawn(player);
         Shop.getPlugin().getLogger().trace("[ShopCreationUtil.sendCreationSuccess] updateSign");
         shop.setSignLinesRequireRefresh(true);
         shop.updateSign();
